requirement.txt
# Unified requirements for running the full project (notebooks + frontend)
# Recommended Python version: 3.10 or 3.11
# For strict legacy reproducibility on Python 3.7, use requirements_legacy_py37.txt

# Core deep learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
tokenizers>=0.13.3

# Optional legacy module compatibility (used by some older code paths)
fastai>=2.7.12

# App/frontend
streamlit>=1.28.0

# Data science stack
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.10.0
tqdm>=4.62.0

# Visualization and image handling
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities used by preprocessing/tokenization and scripts
python-dateutil>=2.8.0
six>=1.16.0
regex>=2022.0.0
requests>=2.28.0
requests-oauthlib>=1.3.0
urllib3>=1.26.0
